Updating pipeline
