+++ qanta/pipeline/dan.py
@@ -7,7 +7,7 @@ from qanta.util import environment as e
 from qanta.guesser.util.format_dan import preprocess
 from qanta.guesser.util import load_embeddings
 from qanta.guesser import dan
-from qanta.guesser.classify.learn_classifiers import print_recall_at_n
+from qanta.guesser.classify.learn_classifiers import print_recall_at_n, print_recall_dan_gpu_at_n
 from qanta.extract_features import create_guesses
 
 
@@ -87,12 +87,12 @@ class EvaluateClassifier(luigi.Task):
     def output(self):
         return LocalTarget(c.EVAL_RES_TARGET)
 
-class EvaluateClassifier(luigi.Task):
+class EvaluateDAN(luigi.Task):
     def requires(self):
-        yield TrainClassifier()
+        yield TrainDAN()
 
     def run(self):
-        print_recall_at_n()
+        print_recall_dan_gpu_at_n()
         
     def output(self):
         return LocalTarget(c.EVAL_RES_TARGET)
